[쇼핑 주문 3, 4단계] 비비 미션 제출합니다. (#124)
* docs: step3 기능명세서 작성

* refactor: RetrofitService 싱글톤으로 변경

* refactor: ProductRepository의 API 호출 로직을 코루틴 적용하여 비동기 처리 개선

* refactor: CartRepository의 fetchAllCart, findCartByProductId, fetchCart 메서드를 suspend 함수로 변경

* refactor: CartRepository의 addToCart 메서드를 suspend 함수로 변경

* refactor: CartRepository의 deleteCart 메서드를 suspend 함수로 변경

* refactor: CartRepository의 updateCart 메서드를 suspend 함수로 변경

* refactor: CartRepository의 getCartCounts 메서드를 suspend 함수로 변경

* refactor: HistoryDao 함수들을 suspend 함수로 변경

* refactor: viewModelScope 불필요한 중첩 제거

* docs: step3 구현 완료 체크

* docs: step4 기능명세서 작성

* feat: 결제 화면 UI 구현

* feat: 주문 버튼 클릭 시 결제 화면으로 이동 기능 구현

* refactor: Product, Cart Client 통합

- ProductClient와 CartClient를 NetworkClient로 통합

* feat: 쿠폰 목록 API 연동 코드 작성

* feat: 쿠폰 데이터 클래스 생성

* feat: 결제 화면 쿠폰 목록 아이템 레이아웃 및 어댑터 구현

* refactor: 쿠폰 목록 API 응답 형식 변경

- 쿠폰 목록을 가져오는 API의 응답 형식을 `ArrayList<CouponResponseItem>` 에서 `List<CouponResponse>`로 변경

* feat: 결제 화면 쿠폰 목록 조회 기능 구현

* feat: 추천 상품 담기/빼기 기능 구현

* feat: 주문할 상품 목록을 결제 화면으로 전달하는 기능 구현

* feat: 주문 금액 정보 표시 기능 구현

* feat: 주문 금액과 시간에 따른 사용 가능 쿠폰 조회 기능 구현

* feat: 결제 화면에서 쿠폰 단일 선택 기능 추가

* feat: DiscountType enum 클래스 구현

* feat: 쿠폰별 사용 가능 조건 검증 기능 구현

* refactor: 쿠폰 적용 로직 개선 및 쿠폰 종류별 구현 분리

* docs: step4 구현 완료 체크

* feat: 주문 API 연동 및 관련 클래스 추가

* feat: 주문 완료 시 상품 추천 화면으로 전환하는 기능 구현

* feat: 주문 완료 시 상품 목록에서 주문한 상품들의 장바구니 수량 0으로 초기화 기능 구현

* fix: 결제 예정 금액 계산 시 배송비 추가

* refactor: 기본 배송비와 할인 금액 상수화

* docs: step4 구현 완료 체크

* refactor: 데이터베이스 이름 상수로 추출

* refactor: Retrofit `@Header` 제거

* fix: CartProduct의 totalPrice 자료형을 Int에서 Long으로 변경

* test: GoodsActivity UI 테스트 코드 수정

* test: ViewModel 테스트 리팩토링

* fix: 추천 상품 추가 시 금액이 정상적으로 반영되도록 수정

* fix: 무료 배송 쿠폰 적용 시 할인 금액 0원으로 수정

* fix: 쿠폰 적용 시 배송비 누락되는 오류 수정

* fix: 추천 상품 수량 변경 시 깜빡임 현상 수정

* test: ViewModel 테스트 코드 추가 및 수정

* refactor: HeaderInterceptor 위치 변경

* refactor: Repository에서 NetworkClient 직접 참조 대신 의존성 주입

* refactor: CouponContract 클래스명을 CouponRule로 변경

* refactor: 쿠폰 선택 상태를 나타내는 변수명 변경 (isChecked -> isApplied)

* refactor: BuyTwoGetOneCoupon을 BuyXGetYCoupon으로 이름 변경

* refactor: 쿠폰 할인 타입 String에서 DiscountType enum으로 변경

* refactor: 가격 정보 기본값 상수화

* refactor: 마지막에 본 상품이 없는 경우에 해당 뷰가 보이지 않도록 수정

* refactor: 주문 요청 시 Cart ID 목록을 전달하도록 변경

* refactor: 장바구니를 하나씩 체크할 때 전체가 체크되었으면 전체 선택 박스도 체크되도록 수정

* refactor: 기본 배송비, 기본 할인 상수 위치 변경

* refactor: 쿠폰 도메인 객체 구조 개선 및 CouponRule 제거

* refactor: 장바구니 실패 상태 세분화 및 UI 반영

* refactor: PaymentViewModel에서 orderIds를 생성자로 전달

* refactor: GoodsDetailsViewModel에서 id를 생성자로 전달

+++ app/src/main/java/woowacourse/shopping/domain/model/CartProduct.kt
@@ -6,7 +6,7 @@ data class CartProduct(
     val quantity: Int,
     val isChecked: Boolean = false,
 ) {
-    val totalPrice: Int get() = product.price * quantity
+    val totalPrice: Long get() = product.price.toLong() * quantity
 
     companion object {
         val EMPTY_CART_PRODUCT =
